requirements.txt
asgiref==3.2.10
autopep8==1.5.4
coverage==5.3
Django==3.1.1
djangorestframework==3.12.1
flake8==3.8.4
mccabe==0.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
wheel==0.35.1
